concurrent_js_update加上判断

+++ jd.sh
@@ -559,30 +559,32 @@ concurrent_js() {
 }
 
 concurrent_js_update() {
-	rm -rf $ccr_js_file/*
-	js_cookie=$(cat $openwrt_script_config/jdCookie.js | sed -e "s/pt_key=XXX;pt_pin=XXX//g" -e "s/pt_pin=(//g" -e "s/pt_key=xxx;pt_pin=xxx//g"| grep "pt_pin" | grep -v "//'" |grep -v "// '")
-	js_amount=$(echo "$js_cookie" |wc -l)
+	if [ "$ccr_if" == "yes" ];then
+		rm -rf $ccr_js_file/*
+		js_cookie=$(cat $openwrt_script_config/jdCookie.js | sed -e "s/pt_key=XXX;pt_pin=XXX//g" -e "s/pt_pin=(//g" -e "s/pt_key=xxx;pt_pin=xxx//g"| grep "pt_pin" | grep -v "//'" |grep -v "// '")
+		js_amount=$(echo "$js_cookie" |wc -l)
 
-	while [[ ${js_amount} -gt 0 ]]; do
-		mkdir $ccr_js_file/js_$js_amount
-		cp $script_dir/jdCookie.js $ccr_js_file/js_$js_amount/jdCookie.js
+		while [[ ${js_amount} -gt 0 ]]; do
+			mkdir $ccr_js_file/js_$js_amount
+			cp $script_dir/jdCookie.js $ccr_js_file/js_$js_amount/jdCookie.js
 
-		if [ ! -L "$ccr_js_file/js_$js_amount/sendNotify.js" ]; then
-			rm -rf $$ccr_js_file/js_$js_amount/sendNotify.js
-			ln -s $script_dir/sendNotify.js $ccr_js_file/js_$js_amount/sendNotify.js
-		fi
+			if [ ! -L "$ccr_js_file/js_$js_amount/sendNotify.js" ]; then
+				rm -rf $$ccr_js_file/js_$js_amount/sendNotify.js
+				ln -s $script_dir/sendNotify.js $ccr_js_file/js_$js_amount/sendNotify.js
+			fi
 
-		js_cookie_obtain=$(echo "$js_cookie" | awk -v a="$js_amount" 'NR==a{ print $0}') #获取pt
-		sed -i '/pt_pin/d' $ccr_js_file/js_$js_amount/jdCookie.js >/dev/null 2>&1
-		sed -i "5a $js_cookie_obtain" $ccr_js_file/js_$js_amount/jdCookie.js
+			js_cookie_obtain=$(echo "$js_cookie" | awk -v a="$js_amount" 'NR==a{ print $0}') #获取pt
+			sed -i '/pt_pin/d' $ccr_js_file/js_$js_amount/jdCookie.js >/dev/null 2>&1
+			sed -i "5a $js_cookie_obtain" $ccr_js_file/js_$js_amount/jdCookie.js
 
-		for i in `ls $dir_file_js | grep -v 'jdCookie.js\|sendNotify.js'`
-		do
-			cp $dir_file_js/$i $ccr_js_file/js_$js_amount/$i
+			for i in `ls $dir_file_js | grep -v 'jdCookie.js\|sendNotify.js'`
+			do
+				cp $dir_file_js/$i $ccr_js_file/js_$js_amount/$i
+			done
+
+			js_amount=$(($js_amount - 1))
 		done
-
-		js_amount=$(($js_amount - 1))
-	done
+	fi
 }
 
 concurrent_js_clean(){
